requirements.txt
Flask-Cors==5.0.0
Flask==3.1.0
IPython
Jinja2==3.1.4
MarkupSafe==3.0.2
Pillow
PyYAML==6.0.2
Pygments==2.18.0
Send2Trash==1.8.3
Werkzeug==3.1.3
accelerate==1.2.1
adjustText==1.3.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
blis==1.0.2
bs4==0.0.2
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.10
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0.tar.gz
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.55.3
fqdn==1.5.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
imageio==2.36.1
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jieba==0.42.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema-specifications==2024.10.1
jsonschema==4.23.0
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
matplotlib==3.10.0
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
murmurhash==1.0.11
narwhals==1.18.4
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk
notebook_shim==0.2.4
numpy==2.0.2
overrides==7.7.0
packaging==24.2
pandas
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.0.0
pipeline
platformdirs==4.3.6
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.0
pytz==2024.2
pywin32==308
pywinpty==2.0.14
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rouge-score
rpds-py==0.22.3
safetensors==0.4.5
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.3.1
sentence_transformers
sentencepiece
sentencepiece==0.2.0
setuptools==75.6.0
shap==0.46.0
shellingham==1.5.4
six==1.17.0
sklearn.model_selection
slicer==0.0.8
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy==3.8.3
spacy>=3.0.0
sqlite3
srsly==2.5.0
stack-data==0.6.3
sympy==1.13.1
terminado==0.18.1
thinc==8.3.2
threadpoolctl==3.5.0
tifffile==2024.12.12
tinycss2==1.4.0
tokenizers==0.21.0
torch==2.5.1
torchmetrics
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers
transformers==4.47.0
trulens
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.17.0